Adding a 2nd property to check on PrintTokens2
