--- java/java-ranger-regression/printtokens_eqchk.yml
@@ -1,10 +1,0 @@
-format_version: "2.0"
-input_files:
-  - ../common/
-  - printtokens_eqchk/
-properties:
-  - property_file: ../properties/assert.prp
-    expected_verdict: true
-
-options:
-  language: Java
