+++ java/java-ranger-regression/printtokens_eqchk/printtokens_prop1.yml
@@ -1,0 +1,9 @@
+format_version: "1.0"
+input_files:
+  - ../../common/
+  - impl/
+  - prop1/
+properties:
+  - property_file: ../../properties/assert.prp
+    expected_verdict: true
+
